+++ test/usecases/repo_contents_usecases_test.dart
@@ -199,6 +199,11 @@ class MockRepoRepository implements RepoRepository {
   Future<Either<Failure, SearchResults>> searchRepos({
     String? q,
     int? uid,
+    String? sort,
+    String? order,
+    bool? private,
+    bool? archived,
+    bool? template,
     int? page,
     int? limit,
   }) => throw UnimplementedError();